requirements.txt
django
djangorestframework
drf_yasg
psycopg2_binary
pillow 
python-slugify
djangorestframework-simplejwt
python-decouple
folium
leaflet
polyline
beautifulsoup4
requests
lxml
